requirements.txt
Brotli==1.0.9
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dash==2.3.0
dash-bootstrap-components==1.0.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.3
Flask-Compress==1.11
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.0
mariadb==1.0.10
MarkupSafe==2.1.1
pdpyras==4.5.0
plotly==5.6.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.32
tenacity==8.0.1
urllib3==1.26.9
Werkzeug==2.0.3
